requirements.txt
langchain==0.1.7
langchain_community==0.0.20
langchain_core==0.1.23
openai==0.28.0
python-dotenv==1.0.1
streamlit==1.30.0
tiktoken==0.5.2
